sptek jpa procedure 테스트

+++ sptek/src/main/java/com/sptek/demo/member/repository/MemberRepository.java
@@ -2,16 +2,15 @@ package com.sptek.demo.member.repository;
 
 import java.util.List;
 
+import com.sptek.demo.member.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sptek.demo.member.model.Member;
-
-public interface MemberRepository extends JpaRepository<Member, Long>{
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 
 	public List<Member> findById(String id);
-	
+
 	public List<Member> findByName(String name);
-	
+
 	public List<Member> findByNameLike(String name);
-	
+
 }
